Testing pause on timers.
There's no built in way to pause a timer, stopping a timer will clear it's current time. However, the remaining time can be captured before and the timings interval can be set to that value before stopping the timer. When started again it will start from the set interval, thus creating a pause feature.